Speed up compilation with precompiled static libs
Added macro definition for making static libraries in makefile (not
documented in comments yet).

Adding static libraries to the repo so that users using this framework
do not need to recompile the static libraries each time they want to
build a project.

Time to run and build all examples projects has gone from 1m25s -> 12s.
