+++ .gitignore
@@ -1,6 +1,7 @@
 firmware/*
 firmware/**/build
 firmware/**/makefile
+firmware/library/static_libraries
 
 !firmware/library
 !firmware/HelloWorld
