+++ firmware/library/L3_Application/L3.mk
@@ -1,5 +1,6 @@
 INCLUDES +=
 SYSTEM_INCLUDES +=
+
 SOURCES += $(LIB_DIR)/L3_Application/task_scheduler.cpp
 
 TESTS += $(LIB_DIR)/L3_Application/task_scheduler.cpp
