+++ firmware/library/newlib/newlib.cpp
@@ -3,6 +3,7 @@
 #include <sys/stat.h>
 #include <unistd.h>
 
+#include <cstdarg>
 #include <cstdint>
 #include <cstdio>
 
@@ -158,3 +159,17 @@ extern "C"
     return i;
   }
 }
+
+#pragma GCC diagnostic ignored "-Wformat-nonliteral"
+// The newlib nano version of scanf relies on malloc.
+// Overriding scanf with an efficient static memory variant.
+// NOLINTNEXTLINE(readability-identifier-naming)
+int scanf(const char * format, ...)
+{
+  va_list args;
+  va_start(args, format);
+  int items = vfscanf(stdin, format, args);
+  va_end(args);
+  return items;
+}
+#pragma GCC diagnostic warning "-Wformat-nonliteral"
